Add initial implementation of real Slack Events handler
This includes the middleware for pulling the Heroku router's X-Request-ID
header, and the middleware for doing the signature and token validation of
inbound requests from Slack.

This also implements the initial Slack Events handler, including doing URL
verification and writing events to the workqueue / sink. The actual writes are
disabled, and instead replaced with debug logging.

This has not really been tested.