Unlearn replying to thread_broadcast messages with its own broadcast
The Go Slack client is missing the field on the MessageEvent that would allow us
to identify who the sender of the message was. Specifically, the `root` key as
shown here: https://api.slack.com/events/message/thread_broadcast

Until that can be done, this disables the bot's ability to reply to broadcasted
messages with a broadcast.